`__init__.py` cleanup
To be able to easily reason about top-level exported objects, make
`__init__.py` clean. Note that to keep compatibility, everything present
in the file before the change was re-exported.
