+++ rohmu/common/constants.py
@@ -1,0 +1,3 @@
+# Copyright (c) 2023 Aiven, Helsinki, Finland. https://aiven.io/
+
+IO_BLOCK_SIZE = 2**20  # 1 MiB
