+++ test/test_encryptor.py
@@ -5,7 +5,7 @@ See LICENSE for details
 from __future__ import annotations
 
 from py.path import LocalPath  # type: ignore [import] # pylint: disable=import-error
-from rohmu import IO_BLOCK_SIZE
+from rohmu.common.constants import IO_BLOCK_SIZE
 from rohmu.encryptor import Decryptor, DecryptorFile, Encryptor, EncryptorFile, EncryptorStream
 from typing import cast, IO
